Option to insert new window next to current window
When I use emamux:new-window (or, emamux:clone-current-frame), I use it
because I want a window which is like a "fork" of current window.  When
using these command in this way, it is more sensible, I believe, to have
the new window next to the current window.

So, making "-a" option available to tmux "new-window" command via prefix
arg.