emamux:run-in-pane to run command in any pane
This is based on issue #45

`emamux:send-command` has the functionality to interactively send
commands to any tmux pane. `emamux:run-command` has the functionality
send commands either interactively or programmatically to an
associated runner pane.

The new function `emamux:run-in-pane` combines parts of both and helps
to send command either interactively or programmatically to any tmux
pane.